Added polynomial benches
